# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Boite \u00e0 id\u00e9e",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Inscription",
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Le nom de l'utilisateur",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "body",
                        "description": "L'email de l'utilisateur",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "body",
                        "description": "Le mot de passe de l'utilisateur",
                        "required": true,
                        "schema": {
                            "properties": {
                                "password": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne les donn\u00e9es de l'utilisateur",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "default": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Connexion",
                "parameters": [
                    {
                        "name": "email",
                        "in": "body",
                        "description": "L'email de l'utilisateur",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "body",
                        "description": "Le mot de passe de l'utilisateur",
                        "required": true,
                        "schema": {
                            "properties": {
                                "password": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "default": "success"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "R\u00e9cup\u00e9rer les donn\u00e9es de l'utilisateur",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "properties": {
                                "bearer": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne les donn\u00e9es de l'utilisateur",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "default": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/ideas": {
            "post": {
                "tags": [
                    "ideas"
                ],
                "summary": "Ajouter une id\u00e9e",
                "parameters": [
                    {
                        "name": "text",
                        "in": "body",
                        "description": "Le texte de l'id\u00e9e",
                        "required": true,
                        "schema": {
                            "properties": {
                                "text": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "id_project",
                        "in": "body",
                        "description": "L'id du projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id_project": {
                                    "type": "integer",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "integer"
                    },
                    {
                        "name": "id_creator",
                        "in": "body",
                        "description": "L'id de l'utilisateur qui ajoute le projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id_creator": {
                                    "type": "integer",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne le projet avec un id"
                    }
                }
            }
        },
        "/ideas/{id}": {
            "put": {
                "tags": [
                    "ideas"
                ],
                "summary": "Modifier une id\u00e9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de l'id\u00e9e",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "integer"
                    },
                    {
                        "name": "text",
                        "in": "body",
                        "description": "Le texte de l'id\u00e9e",
                        "required": true,
                        "schema": {
                            "properties": {
                                "text": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ideas"
                ],
                "summary": "Supprimer une id\u00e9e via son id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id de l'id\u00e9e \u00e0 supprimer",
                        "required": true,
                        "type": "integer",
                        "minimum": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Id\u00e9e supprim\u00e9 avec succ\u00e8s"
                    },
                    "404": {
                        "description": "Id\u00e9e introuvable"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "required": [
                "name",
                "email",
                "password"
            ],
            "properties": {
                "id": {
                    "description": "l'id' du user",
                    "type": "integer"
                },
                "name": {
                    "description": "le nom du user",
                    "type": "string"
                },
                "email": {
                    "description": "l'email du user",
                    "type": "string"
                },
                "password": {
                    "description": "le mot de passe du user",
                    "type": "integer"
                }
            }
        },
...
    },
...
}